+++ R/plotKaplanMeier.R
@@ -25,6 +25,8 @@
 #' @inheritParams ggpubr::ggpar
 #' @inheritDotParams survminer::ggsurvplot
 #' @param surv.median.line.legend Text to be used for median survival line in the legend
+#' @param legend.nrow The number if rows in the legend for the stratification variable
+#' @param legend.ncol The number if columns in the legend for the stratification variable
 #' @details The function takes a dataframe (tibble will be coerced to data.frame) and fits the survival curves using surv_fit then plots them using ggsurvplot.
 #' The arguments risk.table.fontsize and pval.size convert regular sizes to ggplot sizes. For other font size arguments passed to ggsurvplot you can multiply regular point sizes by 0.36 to convert to ggplot sizes. To change the color of the curves use the argument palette instead of color.
 #' @return A "ggsurvplot" object, see the documentation of `survminer::ggsurvplot` for more information. Basically it is a list that notably contains two "ggplot" objects, "x$plot" and "x$table", that renders into a single figure the Kaplan-Meier curves on top and the risk table on the bottom.
@@ -119,10 +121,11 @@ plotKaplanMeier <- function(data,
                             surv.median.line.legend = TRUE,
                             surv.median.line.title  = 'Median Survival',
                             legend                  = "top",
-                            legend.title            = 'Survival curves',
+                            legend.title            = 'Observed',
                             legend.nrow             = NULL,
                             legend.ncol             = NULL,
                             risk.table.ylab         = cov_col,
+                            risk.table.title        ="Number of subjects at risk",
                             legend.labs             = sub(pattern=paste(cov_col,'=', sep = ''),
                                                           replacement="",
                                                           x=names(fit$strata)),
@@ -132,6 +135,8 @@ plotKaplanMeier <- function(data,
                             legend.box              = 'horizontal',
                             legend.title.size       = 9,
                             legend.spacing           = 5,
+                            censor.shape = '|',
+                            censor.size = 4,
                             ...){
   # Check if the input is a dataframe
   if (!is.data.frame(data)){
@@ -170,9 +175,9 @@ plotKaplanMeier <- function(data,
     # if (conf.int.legend == 'Confidence intervals'){
     #   conf.int.legend <- 'Confidence interval'
     # }
-    if (legend.title == 'Survival curves'){
-      legend.title <- 'Survival curve'
-    }
+    # if (legend.title == 'Survival curves'){
+    #   legend.title <- 'Survival curve'
+    # }
   }
 
   #fit survival curves
@@ -181,7 +186,7 @@ plotKaplanMeier <- function(data,
 
   # title for confidence interval legend
   if (add.ciWidth.to.legend){
-    legend.title <- paste0(legend.title,' & ', ciWidth*100, '% CI')
+    legend.title <- paste0(legend.title,' and ', ciWidth*100, '% CI')
   }
   if (!is.null(facet.by)){
     facetPlots <- survminer::ggsurvplot_facet(fit                    = fit,
@@ -206,6 +211,7 @@ plotKaplanMeier <- function(data,
                                               ylim                   = ylim,
                                               palette                = palette,
                                               pval.size              = pval.size * 0.36,
+                                              censor.shape = censor.shape,
                                               ...
     )
     if (surv.median.line %in% c('hv', 'v', 'h')){
@@ -232,16 +238,30 @@ plotKaplanMeier <- function(data,
         )
       }}
 
-        facetPlots <- facetPlots +
-          guides(color= guide_legend(order=1,
-                                     nrow = legend.nrow,
-                                     ncol = legend.ncol
-                                     ),
-                 fill = guide_legend(order=1,
-                                     nrow = legend.nrow,
-                                     ncol = legend.ncol
-                                     ) ,
-                 linetype = guide_legend(order = 3))
+    # Add a blank geom_point to trigger the Censored entry
+    facetPlots <- facetPlots +
+      geom_point(x=NA,y=NA,aes(shape= factor(1))) +
+      scale_shape_manual(name = '',
+                         values = censor.shape,
+                         label = "Censored"
+      )
+
+
+    facetPlots <- facetPlots +
+      guides(color= guide_legend(order=1,
+                                 nrow = legend.nrow,
+                                 ncol = legend.ncol
+      ),
+      fill = guide_legend(order=1,
+                          nrow = legend.nrow,
+                          ncol = legend.ncol,
+                          override.aes=list(shape=NA)
+      ) ,
+      linetype = guide_legend(order = 3),
+      shape = guide_legend(order=3,
+                           override.aes=list(shape=censor.shape,
+                                             linetype=NA,
+                                             size=censor.size)))
 
 
     if (label.parsed == TRUE) {
@@ -308,6 +328,8 @@ plotKaplanMeier <- function(data,
                                            palette               = palette,
                                            pval.size             = pval.size * 0.36,
                                           conf.int.fill          = conf.int.fill,
+                                      censor.shape = censor.shape,
+                                      risk.table.title = risk.table.title,
                                            ...
                                            #font.x               = 25,
                                            #font.y        = 25,
@@ -351,7 +373,16 @@ plotKaplanMeier <- function(data,
       )
     }
 
+    # Add a blank geom_point to trigger the Censored entry
+    kaplanPlot$plot <- kaplanPlot$plot +
+      geom_point(x=NA,y=NA,aes(shape= factor(1))) +
+      scale_shape_manual(name = '',
+                         values = censor.shape,
+                         label = "Censored"
+      )
+
   }
+
   kaplanPlot$plot <- kaplanPlot$plot +
     guides(color= guide_legend(order=1,
                                nrow = legend.nrow,
@@ -359,9 +390,13 @@ plotKaplanMeier <- function(data,
                       ),
              fill = guide_legend(order=1,
                                  nrow = legend.nrow,
-                                 ncol = legend.ncol
+                                 ncol = legend.ncol,
+                                 override.aes = list(shape=NA)
                                ) ,
-           linetype = guide_legend(order = 3)) +
+           linetype = guide_legend(order = 3),
+           shape = guide_legend(order=3,override.aes=list(shape=censor.shape,
+                                                          linetype=NA,
+                                                          size=censor.size))) +
     theme(legend.box = legend.box,
           legend.title = element_text(size= legend.title.size),
           legend.spacing = unit(legend.spacing, 'pt'))