+++ R/summaryFollowUpTime.R
@@ -16,8 +16,10 @@
 #' @param digits_rate the number of significant digits for the Event rate column
 #' @param digits the number of significant digits for the follow up time column
 #' @param caption is the table caption. Assign NULL to this argument produce table without caption.
-#' @param footnote is the text for footnote.
-#' @param label is the label for the table used for cross-reference.
+#' @param footnote is the text for footnote, default is "\\textsuperscript{a}Number of events\\newline\\textsuperscript{b}Total
+#' observation period calculated as sum of individual observation period per row.
+#' \\newline\\textsuperscript{c}Annual event rate calculated, for each row, as number of events per
+#' group divided by total observation period per group."#' @param label is the label for the table used for cross-reference.
 #' @param asList if \strong{TRUE} the data summary is returned as a list where
 #'   different elements in the list contain summary of different stratification
 #'   levels. If \strong{FALSE} the data summary is returned as a latex table. If
@@ -66,9 +68,9 @@ summaryFollowUpTime <- function (df,
                                  caption = "Summary of observation periods and annual event rates in the analysis data set.",
                                  label = "tab:anaSummaryFollowupTime",
                                  footnote = "\\textsuperscript{a}Number of events\\newline\\textsuperscript{b}Total
-                                 observation period calculated as sum of individual observation period per group
-                                 \\newline\\textsuperscript{c}Mean annual event rate calculated as number of events per
-                                 group divided by total observation period per group",
+                                 observation period calculated as sum of individual observation period per row.
+                                 \\newline\\textsuperscript{c}Annual event rate calculated, for each row, as number of events per
+                                 group divided by total observation period per group.",
                                  asList = FALSE,
                                  ...) {